Organize demo
