--- src/Demo/App.xaml.cs
@@ -1,15 +1,0 @@
-﻿namespace UnlockingFlyout
-{
-    public partial class App : Application
-    {
-        public App()
-        {
-            InitializeComponent();
-        }
-
-        protected override Window CreateWindow(IActivationState? activationState)
-        {
-            return new Window(new AppShell());
-        }
-    }
-}
