--- src/Demo/AppShell.xaml.cs
@@ -1,11 +1,0 @@
-﻿namespace UnlockingFlyout
-{
-    public partial class AppShell : Shell
-    {
-        public AppShell()
-        {
-            InitializeComponent();
-			Routing.RegisterRoute(nameof(SecondPage), typeof(SecondPage));
-        }
-    }
-}
