+++ src/Demo/DigitalProduction.Demo.csproj
@@ -65,7 +65,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <MauiXaml Update="SecondPage.xaml">
+	  <MauiXaml Update="Pages\SecondPage.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	</ItemGroup>
