+++ src/Demo/Pages/SecondPage.xaml.cs
@@ -1,0 +1,14 @@
+﻿namespace UnlockingFlyout;
+
+public partial class SecondPage : ContentPage
+{
+	public SecondPage()
+	{
+		InitializeComponent();
+	}
+
+	void OnNavigateBack(object sender, EventArgs eventArgs)
+	{
+		Shell.Current.GoToAsync("../");
+	}
+}
