+++ src/Demo/Program/App.xaml.cs
@@ -1,0 +1,14 @@
+﻿namespace UnlockingFlyout;
+
+public partial class App : Application
+{
+	public App()
+	{
+		InitializeComponent();
+	}
+
+	protected override Window CreateWindow(IActivationState? activationState)
+	{
+		return new Window(new AppShell());
+	}
+}
