+++ src/Demo/Program/AppShell.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace UnlockingFlyout;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+		Routing.RegisterRoute(nameof(SecondPage), typeof(SecondPage));
+	}
+}
