--- src/Demo/SecondPage.xaml.cs
@@ -1,15 +1,0 @@
-﻿namespace UnlockingFlyout
-{
-	public partial class SecondPage : ContentPage
-	{
-		public SecondPage()
-		{
-			InitializeComponent();
-		}
-
-		void OnNavigateBack(object sender, EventArgs eventArgs)
-		{
-			Shell.Current.GoToAsync("../");
-		}
-	}
-}
